Deletes the repo from the pre-commit workflow since windows can't even clone a repo
