+++ .pre-commit-config.yaml
@@ -1,13 +1,7 @@
 repos:
--   repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.6.0
-    hooks:
-    -   id: end-of-file-fixer
-    -   id: check-yaml
-    -   id: check-added-large-files
 -   repo: local
     hooks:
       - id: yapf-diff
         name: yapf-diff
-        entry: python yapf-diff-format.py
+        entry: python pre_commit_runner.py --hook yapf-diff
         language: python
